Hi blocks validator
